backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-multipart==0.0.6
python-dotenv==1.0.0 